Rename worker health check controller / method
**Why**: To be more rails-y
